New counter implementation

+++ LunarMooner/src/PlayerProfile.cpp
@@ -372,7 +372,7 @@ bool PlayerProfile::hasAchievement(AchievementID id) const
 int PlayerProfile::getRank() const
 {
     int i = 0;
-    while (i < rankXP.size() && rankXP[i++] < m_XP) {}
+    while (i < static_cast<int>(rankXP.size()) && rankXP[i++] < m_XP) {}
     return std::max(1, i - 1);
 }
